requirements.txt
einops
lightly
lightning-flash
matplotlib
numpy
opencv-python
pandas
Pillow
pytorch-lightning
rasterio
rioxarray
scikit-image
scikit-learn
scipy
shapely
torchmetrics
tqdm
wandb
xarray